Fix searchPhotos handler on GalleryRouter example
Fix searchPhotos route handler not handling case where a reference
to a page number is not passed. In this case, the console.log
statement would print the page number as being 'undefined'.

The fix also helps to clarify that when the optional page reference
is not supplied, the 'page' parameter of the handling function will have
'undefined' as its value.