Fix broken document structure on the beginning of Stack #2 section
Header for the RESTful Applications > Stack 2 section and end text
for Stack 1 section was not being shown due to a (probably) accidentally
added '</code>' tag.
This also fixes the anchor link on the index for the Stack 2 section
not doing anything.

Also fix header hierarchy for all headers inside Stack 2 section.