fix: directory location in the server.js file
There was a small bug with directory location in the server.js file in the Book Library app in this chapter. 

The main issue was when I (and at least one other person) I typed out the code for the `server.js` file in my application, I got an error in the window: Cannot GET / . Also in the dev tools, http://localhost:4711/ returns a 404 Not Found. 

Removing the `..` from these lines fixes the issue. 

More information can be seen on this issue: https://github.com/addyosmani/backbone-fundamentals/issues/716